Move ActorID into lib.rs
