+++ automerge-protocol/src/utility_impls/mod.rs
@@ -1,3 +1,4 @@
+mod actor_id;
 mod opid;
 mod object_id;
 mod key;
